# Optimize str::repeat