@dtolnay 2018-03-04
Nice work! I tried some tests like the following on my computer and see numbers similar to yours.

```rust
#[bench]
fn repeat100(b: &mut test::Bencher) {
    b.iter(|| { S.repeat(100); });
}
```

|`n`|old [ns/iter]|new [ns/iter]|diff [%]|
---|---|---|---
1 | 28 | 28 | 0
2 | 28 | 28 | 0
3 | 29 | 29 | 0
4 | 33 | 31 | −6.1
5 | 37 | 34 | −8.1
7 | 46 | 35 | −24.
10 | 59 | 37 | −37.
100 | 444 | 65 | −85.